# docs/swagger.yaml
basePath: /
definitions:
...
  keerja-backend_internal_dto_request.UpdateCompanyProfileRequest:
    properties:
      awards:
        type: string
      certifications:
        type: string
      core_values:
        type: string
      description:
        type: string
      employee_count:
        minimum: 0
        type: integer
      facebook_url:
        type: string
      founded_year:
        maximum: 2100
        minimum: 1800
        type: integer
      instagram_url:
        type: string
      linkedin_url:
        type: string
      mission:
        type: string
      twitter_url:
        type: string
      vision:
        type: string
      youtube_url:
        type: string
    type: object
...
  keerja-backend_internal_dto_response.CityResponse:
    properties:
      code:
        type: string
      full_name:
        description: e.g., "Kabupaten Bandung Barat"
        type: string
      id:
        type: integer
      name:
        type: string
      postal_codes:
        type: string
      province_id:
        type: integer
      type:
        description: '"Kota" or "Kabupaten"'
        type: string
    type: object
...
  keerja-backend_internal_dto_response.CompanyListResponse:
    properties:
      companies:
        items:
          $ref: '#/definitions/keerja-backend_internal_dto_response.CompanyResponse'
        type: array
    type: object
  keerja-backend_internal_dto_response.CompanyLocationResponse:
    properties:
      city:
        $ref: '#/definitions/keerja-backend_internal_dto_response.CityResponse'
      district:
        $ref: '#/definitions/keerja-backend_internal_dto_response.DistrictResponse'
      full_location:
        description: e.g., "Batujajar, Kabupaten Bandung Barat, Jawa Barat"
        type: string
      province:
        $ref: '#/definitions/keerja-backend_internal_dto_response.ProvinceResponse'
    type: object
  keerja-backend_internal_dto_response.CompanyProfileResponse:
    properties:
      awards:
        type: string
      certifications:
        type: string
      core_values:
        type: string
      description:
        type: string
      employee_count:
        type: integer
      facebook_url:
        type: string
      founded_year:
        type: integer
      id:
        type: integer
      instagram_url:
        type: string
      linkedin_url:
        type: string
      mission:
        type: string
      twitter_url:
        type: string
      updated_at:
        type: string
      vision:
        type: string
      youtube_url:
        type: string
    type: object
  keerja-backend_internal_dto_response.CompanyResponse:
    properties:
      about:
        type: string
      average_rating:
        type: number
      badge_granted:
        description: Whether company has verification badge
        type: boolean
      banner_url:
        type: string
      city:
        type: string
      company_name:
        type: string
      company_size_detail:
        $ref: '#/definitions/keerja-backend_internal_dto_response.MasterCompanySizeResponse'
      company_type:
        type: string
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      followers_count:
        description: Stats
        type: integer
      full_address:
        type: string
      id:
        type: integer
      industry:
        description: Legacy Fields (for backward compatibility)
        type: string
      industry_detail:
        allOf:
        - $ref: '#/definitions/keerja-backend_internal_dto_response.MasterIndustryResponse'
        description: Master Data Relations (new structure)
      is_active:
        type: boolean
      jobs_count:
        type: integer
      location_detail:
        $ref: '#/definitions/keerja-backend_internal_dto_response.CompanyLocationResponse'
      logo_url:
        type: string
      phone:
        type: string
      province:
        type: string
      reviews_count:
        type: integer
      size_category:
        type: string
      slug:
        type: string
      status:
        description: Verification Status Details
        type: string
      updated_at:
        type: string
      uuid:
        type: string
      verified:
        type: boolean
      verified_at:
        type: string
      website_url:
        type: string
    type: object
...
  keerja-backend_internal_dto_response.DistrictResponse:
    properties:
      city_id:
        type: integer
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  keerja-backend_internal_dto_response.MasterCompanySizeResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      label:
        type: string
      max_employees:
        type: integer
      min_employees:
        type: integer
    type: object
  keerja-backend_internal_dto_response.MasterIndustryResponse:
    properties:
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  keerja-backend_internal_dto_response.ProvinceResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  keerja-backend_internal_utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@keerja.com
    name: API Support
  description: Job platform backend API with authentication, job management, applications,
    companies, and push notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Keerja Backend API
  version: "1.0"
paths:
...
  /companies/{id}/follow:
    delete:
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Unfollow a company
      tags:
      - companies
    post:
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Follow a company
      tags:
      - companies
  /companies/{id}/followers:
    get:
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get company followers
      tags:
      - companies
...
  /companies/{id}/profile:
    get:
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keerja-backend_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/keerja-backend_internal_dto_response.CompanyProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get company profile
      tags:
      - companies
    put:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update company profile request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keerja-backend_internal_dto_request.UpdateCompanyProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update company profile
      tags:
      - companies
  /companies/{id}/profile/publish:
    post:
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Publish company profile
      tags:
      - companies
  /companies/{id}/profile/unpublish:
    post:
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Unpublish company profile
      tags:
      - companies
...
  /users/me/companies/following:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/keerja-backend_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/keerja-backend_internal_dto_response.CompanyListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get followed companies
      tags:
      - companies
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"